adding wiring_shift.c (slow, speedup maybe possible)

+++ hardware/fpga/f32c/cores/f32c/wiring_shift.h
@@ -23,13 +23,13 @@ __BEGIN_DECLS
 /*
  * \brief
  */
-extern uint32_t shiftIn( uint32_t ulDataPin, uint32_t ulClockPin, uint32_t ulBitOrder ) ;
+extern uint8_t shiftIn( uint8_t ulDataPin, uint8_t ulClockPin, uint8_t ulBitOrder ) ;
 
 
 /*
  * \brief
  */
-extern void shiftOut( uint32_t ulDataPin, uint32_t ulClockPin, uint32_t ulBitOrder, uint32_t ulVal ) ;
+extern void shiftOut( uint8_t ulDataPin, uint8_t ulClockPin, uint8_t ulBitOrder, uint8_t ulVal ) ;
 
 __END_DECLS
